Introduce UrlHelperInterface and mark UrlHelper as `@final`
In future this will allow consumers here and elsewhere to accept the interface as a dependency instead of the concrete implementation, enabling UrlHelper to become final whilst providing a safer way of mocking and testing in user projects
